pyproject.toml
[project]
name = "protein-design-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "beautifulsoup4>=4.14.3",
    "biopython>=1.86",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "ipython>=9.9.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "langsmith>=0.6.4",
    "mistralai>=1.10.0",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "py3dmol>=2.5.4",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pymupdf4llm>=0.2.9",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "stmol>=0.0.9",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
    "xmltodict>=1.0.2",
]
